requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
jinja2==3.1.2
python-multipart==0.0.6
httpx==0.25.0
httpcore==0.18.0
h11<0.15,>=0.13
python-dotenv==1.0.0
itsdangerous>=2.0.0

